+++ packages/remix-node/sessions/fileStorage.ts
@@ -9,7 +9,7 @@ import type {
 
 import { createSessionStorage } from "../implementations";
 
-interface FileSessionStorageOptions {
+type FileSessionStorageOptions = {
   /**
    * The Cookie used to store the session id on the client, or options used
    * to automatically create one.
@@ -20,7 +20,7 @@ interface FileSessionStorageOptions {
    * The directory to use to store session files.
    */
   dir: string;
-}
+};
 
 /**
  * Creates a SessionStorage that stores session data on a filesystem.